pass test: init get icons
